% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationFunctions.R
\name{pi.r}
\alias{pi.r}
\title{Correlation prediction interval}
\usage{
pi.r(r, n, rep.n = NA, prob.level = 0.95)
}
\arguments{
\item{r}{Original study: Correlation}

\item{n}{Original study: Sample size}

\item{rep.n}{(optional) Replication study: Sample size. If not specified, n is used.}

\item{prob.level}{(optional 0 to 1 value) Probability level desired (0 to 1). If not specified .95 (i.e., 95 percent) will be used.}
}
\value{
The prediction interval and related statistics in list format.
}
\description{
Correlation prediction interval
}
\examples{
pi.r(r=.35,n=100,rep.n=200)
}

